app/src-tauri/agent/requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.21
perplexityai

